refactor: import mergeEndpointData
